change: make the compressor run in a separate thread
Before the change the compressor abuses asyncio to run as much code as
possible in a single thread. Since the compression algorithm is
relatively slow, running the code synchronously might slow down the
program a lot. Even though threading is known to have relatively poor
performance on python, I believe it's better than asyncio in this
situation as there is less overhead in threading than asyncio with
the old code.